feat: cast section까지 완료
